backend/pyproject.toml
[tool.poetry]
name = "parcel-tracking-poc"
version = "0.1.0"
description = ""
authors = ["shweta-pml <shweta@plusminuslabs.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
httpx = "^0.19.0"
six = "^1.16.0"
aftership = "^1.3.0"
lxml = "^4.6.3"
requests = "^2.26.0"
isort = "^5.9.3"
black = "^21.9b0"
mypy = "^0.910"
flake8 = "^4.0.1"
elasticsearch = "^7.15.1"
aiohttp = "^3.7.4"
loguru = "^0.5.3"
beautifulsoup4 = "^4.10.0"
numpy = "1.21.0"
arrow = "^1.2.1"
pycurl = "^7.44.1"
elasticsearch-dsl = "^7.4.0"
pydash = "^5.1.0"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"
sentry-sdk = "^1.4.3"
python-dotenv = "^0.19.1"
aioredis = "^2.0.0"
google-auth = "^2.3.3"
fastapi-sessions = "^0.3.2"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
